feat(insights): structure, data-quality & coverage fallback descriptions
